refactor: introduce shared 'real' type across SelVeri components and move all ast nodes under ast.py
