+++ src/selveri/common/types.py
@@ -1,0 +1,6 @@
+"""Shared scalar types used across SelVeri."""
+
+from __future__ import annotations
+
+# SelVeri REAL values are represented as Python floats in the AST, IR, and VM.
+real = float
